+++ test/plugin/scenarios/clickhouse-0.3.2.x-scenario/src/main/java/org/apache/skywalking/apm/testcase/neo4j/Application.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.apm.testcase.neo4j;
+
+import com.clickhouse.jdbc.ClickHouseDataSource;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+import java.sql.SQLException;
+import java.util.Properties;
+
+@SpringBootApplication
+public class Application {
+
+    @Value("${clickhouse.jdbc.url:jdbc:clickhouse://clickhouse-server:8123/system}")
+    private String clickhouseJdbcUrl;
+
+    public static void main(String[] args) {
+        try {
+            SpringApplication.run(Application.class, args);
+        } catch (Exception e) {
+            // Never do this
+        }
+    }
+
+    @Bean
+    public ClickHouseDataSource dataSource() throws SQLException {
+        Properties properties = new Properties();
+        properties.put("clientName", "Agent #1");
+        properties.put("sessionId", "default-session-id");
+
+        return new ClickHouseDataSource(clickhouseJdbcUrl, properties);
+    }
+}